make DRYRUN_PREFIX a display attribute, move DryRunClient out of pkg/api
